+++ src/components/Auth/index.jsx
@@ -1,0 +1,3 @@
+export * from "./Login";
+export * from "./Register";
+export * from "./Logout";
