+++ src/components/Common/Input.jsx
@@ -1,4 +1,4 @@
-const Input = ({ type, name, placeholder, value, onChange }) => {
+export const Input = ({ type, name, placeholder, value, onChange }) => {
   return (
     <input
       className="input-field" // This class will inherit styles from global CSS
@@ -12,4 +12,3 @@ const Input = ({ type, name, placeholder, value, onChange }) => {
     />
   );
 };
-export default Input;
