+++ src/components/Common/PrivateRoute.jsx
@@ -2,10 +2,8 @@
 import React from "react";
 import { Navigate } from "react-router-dom";
 
-const PrivateRoute = ({ element, ...rest }) => {
+export const PrivateRoute = ({ element, ...rest }) => {
   const isAuthenticated = localStorage.getItem("token");
 
   return isAuthenticated ? element : <Navigate to="/login" replace />;
 };
-
-export default PrivateRoute;
