+++ src/components/Common/Sidebar.jsx
@@ -2,7 +2,12 @@ import React from "react";
 import { useNavigate } from "react-router-dom";
 import "../../css/sidebar.css";
 
-const Sidebar = ({ authenticated, onLogout, links = [], sidebarOpen }) => {
+export const Sidebar = ({
+  authenticated,
+  onLogout,
+  links = [],
+  sidebarOpen,
+}) => {
   const navigate = useNavigate();
 
   const handleNavigation = (name) => {
@@ -27,5 +32,3 @@ const Sidebar = ({ authenticated, onLogout, links = [], sidebarOpen }) => {
     </div>
   );
 };
-
-export default Sidebar;
