+++ src/components/Profile/FollowersList.jsx
@@ -1,34 +1,33 @@
 import React, { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
-import { getUserFollowing, followUser } from "../../apis/profile";
-import FollowButton from "./FollowButton";
+import { FollowButton } from "./index";
 import "./css/userlist.css";
+import { followUser, getUserFollowers } from "../../apis";
+import { DEFAULT_PROFILE_PIC } from "../../constants";
 
-const DEFAULT_PROFILE_PIC = "/assets/profilePic.jpeg";
-
-const FollowingList = () => {
+export const FollowersList = () => {
   const { userId } = useParams();
-  const [following, setFollowing] = useState([]);
+  const [followers, setFollowers] = useState([]);
   const [loggedInUserId] = useState(localStorage.getItem("userId"));
   const [followingStates, setFollowingStates] = useState({});
 
   useEffect(() => {
-    const fetchFollowing = async () => {
+    const fetchFollowers = async () => {
       try {
-        const data = await getUserFollowing(userId);
-        setFollowing(data);
-
+        const data = await getUserFollowers(userId);
+        setFollowers(data);
         const followStatus = {};
-        data.forEach((user) => {
-          followStatus[user.id] = user.followers?.includes(loggedInUserId);
+        data.forEach((follower) => {
+          followStatus[follower.id] =
+            follower.followers?.includes(loggedInUserId);
         });
         setFollowingStates(followStatus);
       } catch (err) {
-        console.error("Error fetching following:", err);
+        console.error("Error fetching followers:", err);
       }
     };
 
-    fetchFollowing();
+    fetchFollowers();
   }, [userId, loggedInUserId]);
 
   const toggleFollow = async (targetUserId) => {
@@ -39,50 +38,48 @@ const FollowingList = () => {
         [targetUserId]: !prev[targetUserId],
       }));
     } catch (err) {
-      console.error("Error toggling follow:", err);
+      console.error("Follow/unfollow failed:", err);
     }
   };
 
   return (
     <div className="modal">
       <div className="modal-content">
-        <h3>Following</h3>
+        <h3>Followers</h3>
         <ul className="user-list">
-          {following.length > 0 ? (
-            following.map((user) => (
-              <li className="user-item" key={user.id}>
+          {followers.length > 0 ? (
+            followers.map((follower) => (
+              <li className="user-item" key={follower.id}>
                 <div className="user-info">
                   <img
                     src={DEFAULT_PROFILE_PIC}
-                    // src={user.profilePic || DEFAULT_PROFILE_PIC}
-                    alt={user.username}
+                    // src={follower.profilePic || DEFAULT_PROFILE_PIC}
+                    alt={follower.username}
                     className="user-avatar"
                   />
                   <div className="user-details">
                     <span className="user-name">
-                      {user.firstName} {user.lastName}
+                      {follower.firstName} {follower.lastName}
                     </span>
-                    <span className="user-handle">@{user.username}</span>
+                    <span className="user-handle">@{follower.username}</span>
                   </div>
                 </div>
 
-                {user.id !== loggedInUserId && (
+                {follower.id !== loggedInUserId && (
                   <div className="follow">
                     <FollowButton
-                      isFollowing={followingStates[user.id]}
-                      toggleFollow={() => toggleFollow(user.id)}
+                      isFollowing={followingStates[follower.id]}
+                      toggleFollow={() => toggleFollow(follower.id)}
                     />
                   </div>
                 )}
               </li>
             ))
           ) : (
-            <p>Not following anyone yet.</p>
+            <p>No followers yet.</p>
           )}
         </ul>
       </div>
     </div>
   );
 };
-
-export default FollowingList;
