+++ src/components/Profile/index.jsx
@@ -1,0 +1,7 @@
+export * from "./CoverPhotoUploader";
+export * from "./ProfilePicUploader";
+export * from "./FollowButton";
+export * from "./FollowersList";
+export * from "./FollowingList";
+export * from "./Profile";
+export * from "./ProfileEdit";
