+++ src/constants/index.jsx
@@ -1,0 +1,2 @@
+export const DEFAULT_COVER_PHOTO = "/assets/backgroundWall.jpg";
+export const DEFAULT_PROFILE_PIC = "/assets/profilePic.jpeg";
